+++ CHANGELOG.md
@@ -18,6 +18,10 @@ This project adheres to [Semantic Versioning](https://semver.org).
 * Neaten the files output with ``--show-files``.
 * Neaten the abbrs output.
 
+### Fixed
+
+* Properly set base path for directory input.
+
 
 ## [2.0.0](https://github.com/thebigmunch/thorod/releases/tag/2.0.0) (2019-10-18)